Added admin Implementation

+++ FootballProjectSoftUni/Controllers/CityController.cs
@@ -1,10 +1,13 @@
 ﻿using FootballProjectSoftUni.Core.Contracts.City;
 using FootballProjectSoftUni.Core.Models.City;
+using FootballProjectSoftUni.Extensions;
 using FootballProjectSoftUni.Infrastructure.Data.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace FootballProjectSoftUni.Controllers
 {
+    [Authorize]
     public class CityController : Controller
     {
         private readonly ICityService cityService;
@@ -14,6 +17,7 @@ namespace FootballProjectSoftUni.Controllers
             cityService = _service;
         }
 
+        [AllowAnonymous]
         [HttpGet]
         public async Task<IActionResult> All(int? page)
         {
@@ -25,6 +29,11 @@ namespace FootballProjectSoftUni.Controllers
         [HttpGet]
         public IActionResult AddCity()
         {
+            if (User.IsAdmin() == false)
+            {
+                return Unauthorized();
+            }
+
             var model = new CityViewModel();
 
             return View(model);
@@ -33,6 +42,11 @@ namespace FootballProjectSoftUni.Controllers
         [HttpPost]
         public async Task<IActionResult> AddCity(CityViewModel model)
         {
+            if (User.IsAdmin() == false)
+            {
+                return Unauthorized();
+            }
+
             if (ModelState.IsValid == false)
             {
                 return View(model);
@@ -47,6 +61,10 @@ namespace FootballProjectSoftUni.Controllers
 
         public async Task<IActionResult> DeleteCity()
         {
+            if (User.IsAdmin() == false)
+            {
+                return Unauthorized();
+            }
 
             var cities = await cityService.AllCitiesAsync(); 
 
@@ -58,6 +76,11 @@ namespace FootballProjectSoftUni.Controllers
 
         public async Task<IActionResult> DeleteConfirmed(CityViewModel model, int id)
         {
+            if (User.IsAdmin() == false)
+            {
+                return Unauthorized();
+            }
+
             var town = await cityService.FindTownAsync(model, id);
 
             if (town == null)
@@ -73,11 +96,17 @@ namespace FootballProjectSoftUni.Controllers
 
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (User.IsAdmin() == false)
+            {
+                return Unauthorized();
+            }
+
             await cityService.DeleteCityAsync(id);
 
             return RedirectToAction(nameof(All));
         }
 
+        [AllowAnonymous]
         [HttpGet]
         public async Task<IActionResult> Search(string searchString)
         {
